src/seed/requirements.txt
pyzmq
pandas
numpy
keras
tensorflow
flwr
